+++ KuleumBridge/app/src/main/res/layout/home_layout.xml
@@ -21,15 +21,7 @@
         android:visibility="visible">
 
 
-        <androidx.appcompat.widget.Toolbar
-            xmlns:android="http://schemas.android.com/apk/res/android"
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
-            android:theme="@style/Theme.NavigationDrawer"
-            android:minHeight="?attr/actionBarSize">
-        </androidx.appcompat.widget.Toolbar>
+        <include layout="@layout/toolbar"/>
 
 
         <TextView
